now using custom xml tree code instead of crap code from 'net
