+++ .vscode/launch.json
@@ -5,12 +5,19 @@
 	"version": "0.2.0",
 	"configurations": [
 		{
-			"name": "Launch",
+			"name": "XML Tree",
 			"type": "go",
 			"request": "launch",
-			"mode": "auto",
-			"program": "${fileDirname}",
-			// "cwd": "${workspaceFolderBasename}"
+			"mode": "debug",
+			"program": "code/cmd/xmltree/main.go",
+			"output": "code/bin/xmltree.exe",
+			"cwd": "${workspaceFolder}",
+			"args": [
+				"-source",
+				"XL/Governments_Dhayut.xml.xml",
+				"-target",
+				"temp/Governments_Dhayut.xml"
+			]
 		},
 		{
 			"name": "Replace Quameno Ship Hulls",
@@ -44,20 +51,5 @@
 				"XL Armor.csv",
 			]
 		},
-		{
-			"name": "XML",
-			"type": "go",
-			"request": "launch",
-			"mode": "debug",
-			"program": "code/cmd/xml/main.go",
-			"output": "code/bin/xml.exe",
-			"cwd": "${workspaceFolder}",
-			"args": [
-				"-source",
-				"XL/ShipHulls_Dhayut.xml",
-				"-target",
-				"temp/ShipHulls_Dhayut.xml"
-			]
-		}
 	]
 }
